refactor: 불필요한 LunchMenuView 객체 삭제 및 LunchMenuController 이름 수정
