--- src/controller/LunchMenuController.js
@@ -1,50 +1,0 @@
-import { restaurants } from '../domain/restaurants';
-
-import LunchMenuView from '../view/LunchMenuView';
-
-import { $ } from '../utils/dom';
-import { setLocalStorage } from '../utils/localStorage';
-
-const LunchMenuController = {
-  init() {
-    restaurants.init();
-    LunchMenuView.render(restaurants.list);
-    LunchMenuView.bindEvents();
-    this.bindEvents();
-  },
-
-  bindEvents() {
-    $('restaurant-register-modal').addEventListener('registerRestaurant', (e) =>
-      this.handleRestaurantRegister(e.detail)
-    );
-    $('.restaurant-filter-container').addEventListener('change', () =>
-      this.handleRestaurantFilter()
-    );
-  },
-
-  handleRestaurantRegister(data) {
-    restaurants.add(data);
-    this.setRestaurantList();
-    this.handleRestaurantFilter();
-    LunchMenuView.closeModal();
-  },
-
-  setRestaurantList() {
-    setLocalStorage('restaurants', restaurants.list);
-  },
-
-  handleRestaurantFilter() {
-    const sortingType = $('#sorting-filter').value;
-    const category = $('#category-filter').value;
-    const filteredRestaurants = restaurants.filterByCategory(category, restaurants.list);
-    const sortByType = {
-      register: () => filteredRestaurants,
-      name: () => restaurants.sortByName(filteredRestaurants),
-      distance: () => restaurants.sortByDistance(filteredRestaurants),
-    };
-
-    LunchMenuView.render(sortByType[sortingType]());
-  },
-};
-
-export default LunchMenuController;
